runtime/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
openai>=2.0.0
qdrant-client>=1.17.0
rank-bm25>=0.2.2
duckdb>=1.0.0
pyarrow>=15.0.0
numpy>=2.0.0
python-dotenv>=1.0.0
boto3>=1.34.0
requests>=2.32.0
